Replace image placeholders with appropriate SVGs
It was decided to download the SVGs and keep them in a folder instead of
directly imbedding them into the HTML, as there's no need to control or
edit them in this project, saving us a bunch of screen space.

The remaining image placeholders for the user icons will be non-SVG
images that will be retrieved in a later update.